requirements.txt
altair==5.5.0
attrs==25.3.0
blinker==1.6.2
brotli==1.0.9
cachetools==6.2.0
certifi==2024.12.5
cffi==1.16.9
charset-normalizer==3.3.0
click==8.1.7
Flask==3.2.2
geographiclib==2.1
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.45
gunicorn==20.1.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.0
idna==3.5
importlib-metadata==8.13.0
itsdangerous==2.2.1
Jinja2==3.1.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==2.1.3
narwhals==2.6.0
numpy==2.3.3
packaging==23.2
pandas==2.3.3
pillow==11.3.0
protobuf==6.32.1
pyarrow==21.0.0
pycparser==2.21
pydeck==0.9.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.31.1
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.1.0
Werkzeug==2.3.8
zipp==3.19.0
zstandard==0.25.0
